settings

+++ app/Services/Dashboard/BrandService.php
@@ -51,7 +51,7 @@ class BrandService
 
     public function createBrand($data)
     {
-        if ($data['logo'] != null) {
+        if (array_key_exists('logo', $data) && $data['logo'] != null) {
             $file_name = $this->imageManger->uploadSingleImage('/', $data['logo'], 'brands');
             $data['logo'] = $file_name;
         }
@@ -64,8 +64,8 @@ class BrandService
     public function updateBrand($id, $data)
     {
         $brand = $this->getBrand($id);
-// dd($data);
-        if (array_key_exists('logo', $data)) {
+        // dd($data);
+        if (array_key_exists('logo', $data) && $data['logo'] != null) {
             // delete old logo
             $this->imageManger->deleteImageFromLocal($brand->logo);
